Ensure twist is the same for all corners
This prevents the order of corner tupless from being reversed after
turning a face, they are only rotated -1, 0, 1 times for a 90 degree
turn on a cube

These are now the only possible changes for the UFR corner are {RUF,
UFR, FRU}

This change allows corner orientation to be mapped to a number allowing
an index and orientation representation for all polyhedra. This is
faster than the tuple model as it more compressed and makes it easier to
check move tables.